requirements.txt
Pillow>=9.0
pyyaml>=5.0
